MicroserviceWebApplication/requirements.txt
Django~=5.1.6
djangorestframework~=3.15.2
requests~=2.32.3
djangorestframework_simplejwt==5.5.0
python-decouple~=3.8
dj-database-url==2.3.0
psycopg2==2.9.10